♻️ refactor : search 페이지 일부를 ssr화. 🚑fix : getUser 함수 수정.
