+++ package.json
@@ -31,7 +31,7 @@
     "openai": "^4.77.4",
     "react": "^18",
     "react-day-picker": "^9.5.0",
-    "react-dom": "^19.0.0",
+    "react-dom": "^18.0.0",
     "react-hook-form": "^7.54.2",
     "react-icons": "^5.4.0",
     "react-select": "^5.9.0",
