+++ src/app/(home)/page.tsx
@@ -1,6 +1,7 @@
 import CategoryPage from './[id]/page';
 
 const Homepage = () => {
+  
   return (
     <>
       <CategoryPage params={{ id: 'All' }} />
