+++ src/utils/api/supabase_api/fetchUserProfile.ts
@@ -1,0 +1,7 @@
+import { supabase } from '@/utils/supabase/supabaseClient';
+import { getUser } from './getUser';
+
+export const fetchUserProfile = async () => {
+  const data = await getUser();
+  return data; // 인증된 유저의 프로필 반환
+};
